refactor: reorganize `remote_function` code for readability
